Create tests folder; move test code from raytracing.py to test.py
